ATS/requirements.txt
google-generativeai
python-dotenv
pypdf2
